backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
psycopg>=3.1.8,<3.2
drf-spectacular>=0.26.5,<0.27